BuySeed&PlantSeed&HarvestFruit
게임매니저 이용해서 씨앗 구매하기 기능 로직 추가

씨앗 심기 선택창 만들고 버튼이랑 연동 & 씨앗 심기 함수랑 연동

과일 수확하기 기능

+++ Assets/02.Scripts/Farm/Seed.cs
@@ -8,7 +8,7 @@ public class Seed : MonoBehaviour
     // 만들어놓은 프리팹은 SeedContainer 에 저장할 것..
 
     public string seedName;
-    public float seedPrice;
+    public int seedPrice;
     public float growTime; // 성장하는데 걸리는 시간
     public float currentTime; // 심은 후부터 현재까지 시간
